+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # makfy changelog
 
+## 1.1.0
+
+- breaking changes:
+  - removed utils.filesChanged
+  - added utils.getFileChangesAsync, a better approach to detect which files have changed
+  - renamed utils.cleanCacheSync to utils.cleanCache
+
 ## 1.0.0
 
 - first release